+++ packages/api-apw/src/storageOperations/commentStorageOperations.ts
@@ -5,6 +5,7 @@ import { getFieldValues, getTransformer } from "~/utils/fieldResolver";
 import WebinyError from "@webiny/error";
 import { ApwComment } from "~/types";
 import { CmsEntryListParams } from "@webiny/api-headless-cms/types";
+import { COMMENT_MODEL_ID } from "~/storageOperations/models/comment.model";
 
 const pickIdFromChangeRequest = (obj: Record<string, any>): ApwComment => {
     const rawValue = obj["changeRequest"];
@@ -17,13 +18,16 @@ const pickIdFromChangeRequest = (obj: Record<string, any>): ApwComment => {
 
 export const createCommentStorageOperations = ({
     cms,
-    getCmsContext
+    getCmsContext,
+    security
 }: CreateApwStorageOperationsParams): ApwCommentStorageOperations => {
     const getCommentModel = async () => {
-        const model = await cms.getModel("apwCommentModelDefinition");
+        security.disableAuthorization();
+        const model = await cms.getModel(COMMENT_MODEL_ID);
+        security.enableAuthorization();
         if (!model) {
             throw new WebinyError(
-                "Could not find `apwContentReviewModelDefinition` model.",
+                `Could not find "${COMMENT_MODEL_ID}" model.`,
                 "MODEL_NOT_FOUND_ERROR"
             );
         }
@@ -31,7 +35,9 @@ export const createCommentStorageOperations = ({
     };
     const getComment: ApwCommentStorageOperations["getComment"] = async ({ id }) => {
         const model = await getCommentModel();
+        security.disableAuthorization();
         const entry = await cms.getEntryById(model, id);
+        security.enableAuthorization();
         return getFieldValues({
             entry,
             fields: baseFields,
@@ -47,10 +53,12 @@ export const createCommentStorageOperations = ({
         },
         async listComments(params) {
             const model = await getCommentModel();
+            security.disableAuthorization();
             const [entries, meta] = await cms.listLatestEntries(
                 model,
                 params as CmsEntryListParams
             );
+            security.enableAuthorization();
             const values = await Promise.all(
                 entries.map(entry =>
                     getFieldValues<ApwComment>({
@@ -67,6 +75,7 @@ export const createCommentStorageOperations = ({
         async createComment(this: ApwStorageOperations, params) {
             const model = await getCommentModel();
             const refModel = await this.getChangeRequestModel();
+            security.disableAuthorization();
             const entry = await cms.createEntry(model, {
                 ...params.data,
                 changeRequest: {
@@ -74,6 +83,7 @@ export const createCommentStorageOperations = ({
                     modelId: refModel.modelId
                 }
             });
+            security.enableAuthorization();
 
             const values = await getFieldValues({
                 entry,
@@ -91,10 +101,12 @@ export const createCommentStorageOperations = ({
              */
             const existingEntry = await getComment({ id: params.id });
 
+            security.disableAuthorization();
             const entry = await cms.updateEntry(model, params.id, {
                 ...existingEntry,
                 ...params.data
             });
+            security.enableAuthorization();
             const values = await getFieldValues({
                 entry,
                 fields: baseFields,
@@ -105,7 +117,9 @@ export const createCommentStorageOperations = ({
         },
         async deleteComment(params) {
             const model = await getCommentModel();
+            security.disableAuthorization();
             await cms.deleteEntry(model, params.id);
+            security.enableAuthorization();
             return true;
         }
     };