+++ packages/app-headless-cms/src/HeadlessCMS.tsx
@@ -1,6 +1,7 @@
 import React, { Fragment, memo } from "react";
 import { ApolloClient } from "apollo-client";
 import { plugins } from "@webiny/plugins";
+import { ApolloCacheObjectIdPlugin } from "@webiny/app";
 import { Plugins, Provider } from "@webiny/app-admin";
 import { CmsProvider } from "~/admin/contexts/Cms";
 import { CmsMenuLoader } from "~/admin/menus/CmsMenuLoader";
@@ -35,6 +36,16 @@ export interface HeadlessCMSProps {
 
 const HeadlessCMSExtension = ({ createApolloClient }: HeadlessCMSProps) => {
     plugins.register(apiInformation);
+
+    plugins.register(
+        new ApolloCacheObjectIdPlugin(obj => {
+            if (obj.__typename === "CmsContentModelField") {
+                return null;
+            }
+
+            return undefined;
+        })
+    );
 
     return (
         <Fragment>