+++ packages/pulumi-aws/src/apps/core/createCorePulumiApp.ts
@@ -59,14 +59,14 @@ export function createCorePulumiApp(projectAppParams: CreateCorePulumiAppParams 
             }
 
             const prod = app.params.run.env === "prod";
-            const protect = app.getParam(projectAppParams.protect) || prod;
+            const protect = app.getParam(projectAppParams.protect) ?? prod;
             const legacyConfig = app.getParam(projectAppParams.legacy) || {};
 
             // Setup DynamoDB table
             const dynamoDbTable = app.addModule(CoreDynamo, { protect });
 
             // Setup VPC
-            const vpcEnabled = app.getParam(projectAppParams?.vpc) || prod;
+            const vpcEnabled = app.getParam(projectAppParams?.vpc) ?? prod;
             const vpc = vpcEnabled ? app.addModule(CoreVpc) : null;
 
             // Setup Cognito