+++ packages/serverless-cms-aws/src/core/plugins/handlersPaths.ts
@@ -1,1 +1,0 @@
-export const HANDLERS_PATHS = [["dynamoToElastic"]];